refactor: Make system_port and system_host options common
