+++ appium/options/mac/mac2/base.py
@@ -21,6 +21,8 @@ from appium.options.common.automation_name_option import AUTOMATION_NAME
 from appium.options.common.base import PLATFORM_NAME, AppiumOptions
 from appium.options.common.postrun_option import PostrunOption
 from appium.options.common.prerun_option import PrerunOption
+from appium.options.common.system_host_option import SystemHostOption
+from appium.options.common.system_port_option import SystemPortOption
 
 from .arguments_option import ArgumentsOption
 from .bootstrap_root_option import BootstrapRootOption
@@ -29,8 +31,6 @@ from .environment_option import EnvironmentOption
 from .server_startup_timeout_option import ServerStartupTimeoutOption
 from .show_server_logs_option import ShowServerLogsOption
 from .skip_app_kill_option import SkipAppKillOption
-from .system_host_option import SystemHostOption
-from .system_port_option import SystemPortOption
 from .web_driver_agent_mac_url_option import WebDriverAgentMacUrlOption
 
 
@@ -49,6 +49,47 @@ class Mac2Options(
     SystemPortOption,
     WebDriverAgentMacUrlOption,
 ):
+    @PrerunOption.prerun.setter  # type: ignore
+    def prerun(self, value: Dict[str, str]) -> None:
+        """
+        A mapping containing either 'script' or 'command' key. The value of
+        each key must be a valid AppleScript script or command to be
+        executed after before Mac2Driver session is started. See
+        https://github.com/appium/appium-mac2-driver#applescript-commands-execution
+        for more details.
+        """
+        PrerunOption.prerun.fset(self, value)  # type: ignore
+
+    @PostrunOption.postrun.setter  # type: ignore
+    def postrun(self, value: Dict[str, str]) -> None:
+        """
+        A mapping containing either script or command key. The value of
+        each key must be a valid AppleScript script or command to be
+        executed after Mac2Driver session is stopped. See
+        https://github.com/appium/appium-mac2-driver#applescript-commands-execution
+        for more details.
+        """
+        PostrunOption.postrun.fset(self, value)  # type: ignore
+
+    @SystemPortOption.system_port.setter
+    def system_port(self, value: int) -> None:
+        """
+        Set the number of the port for the internal server to listen on.
+        If not provided then Mac2Driver will use the default port 10100.
+        """
+        SystemPortOption.system_port.fset(self, value)
+
+    @SystemHostOption.system_host.setter
+    def system_host(self, value: str) -> None:
+        """
+        Set the number of the port for the internal server to listen on.
+        If not provided then Mac2Driver will use the default host
+        address 127.0.0.1. You could set it to 0.0.0.0 to make the
+        server listening on all available network interfaces.
+        It is also possible to set the particular interface name, for example en1.
+        """
+        SystemHostOption.system_host.fset(self, value)
+
     @property
     def default_capabilities(self) -> Dict:
         return {
